requirements.txt
datasets
httpx
python-dotenv
networkx
matplotlib
pandas